fix: address CodeRabbit review findings on #603
Nine of eleven fixed; one already fixed on this branch, one deferred with
reasoning.

Correctness / security

- tray: only clear the encryption warning on a POSITIVE result. The probe
  `is_plaintext_sqlite` answers `false` both for "confirmed encrypted" and
  for "could not read the file", so an unreadable database silently
  dismissed a warning that was still true. Added `db_crypto::plaintext_state`
  (three-valued) and made the notice decision Raise / Clear / Leave;
  `is_plaintext_sqlite` keeps its exact old semantics for its own callers,
  where collapsing None to false is correct.
- db_crypto: `finalize_encryption_swap` now logs at ERROR on all three
  failure branches - including the one where rollback also failed and the
  data needs restoring by hand. That branch previously reached telemetry
  only if some caller happened to re-log it, which nothing did.
- tray/account: `write_account_pseudonym` logs its failure boundary. The
  mapped String error stopped at the Tauri command boundary, so a failing
  settings write left the Support ID stale invisibly - the one value support
  uses to find a user's error rows.

Observability

- tray: the encryption-notice path gets a span carrying the chosen action,
  and its failure is ERROR rather than WARN. No db_path on the span - a
  home-dir path is user data.

Quality

- The four providers that hand-rolled classify -> warn -> note/stamp now
  call `record_sync_failure`. That policy lived in five places; a change to
  it had to be made in all of them or silently diverge. jira's auth-resolve
  arm keeps its own copy, commented, because it needs the remedy override
  the shared helper cannot express.
- github: reuse `http::chain` instead of a hand-rolled `{e:#}`.
- Extracted `providers/mod.rs`'s tests to `sync_failure_tests.rs`; the
  module had grown past the 500-line rule (338 + 434 now).
- redact: compute the host pseudonym once per payload instead of once per
  attribute. It reads settings.json, and `host.name` is on every resource,
  scope and record - so one ship was hundreds of file reads. Per-payload
  rather than process-wide on purpose: the account pseudonym changes on
  sign-in/out and the next payload must see it.
- db_crypto: em-dash in a displayed error message.

Not fixed

- Pinning cache-warm.yml's remaining actions to SHAs. Deferred to #605,
  which already covers this and proposes doing release-build.yml (which
  holds the signing key and tokens) in the same sweep rather than hardening
  the lower-privilege workflow alone.
- The em-dash in the tray notice was already fixed earlier on this branch.

821 daemon, 169 tray. clippy -D warnings + fmt clean on both workspaces.
